+++ spec/requests/as_content_owner/channel_spec.rb
@@ -111,6 +111,15 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'views can be grouped by video' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :video' do
+          views = channel.views range.merge by: :video
+          expect(views.keys).to all(be_instance_of Yt::Video)
+        end
+      end
+
       describe 'comments can be retrieved for a specific day' do
         context 'in which the channel was partnered' do
           let(:comments) { channel.comments_on 5.days.ago}
@@ -541,6 +550,15 @@ describe Yt::Channel, :partner do
         specify 'with the :by option set to :traffic_source' do
           estimated_minutes_watched = channel.estimated_minutes_watched range.merge by: :traffic_source
           expect(estimated_minutes_watched.keys - keys).to be_empty
+        end
+      end
+
+      describe 'estimated minutes watched can be grouped by video' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :video' do
+          estimated_minutes_watched = channel.estimated_minutes_watched range.merge by: :video
+          expect(estimated_minutes_watched.keys).to all(be_instance_of Yt::Video)
         end
       end